Updated block components - removed duplicate, added flag as insertable
